+++ Tests/SwiftLintFrameworkTests/RulesTests.swift
@@ -117,6 +117,10 @@ class RulesTests: XCTestCase {
         verifyRule(ExplicitInitRule.description)
     }
 
+    func testFileHeader() {
+        verifyRule(FileHeaderRule.description, commentDoesntViolate: false)
+    }
+
     func testFileLength() {
         verifyRule(FileLengthRule.description, commentDoesntViolate: false)
     }